Separate inference rules to different modules
