+++ crates/analyzer-module/src/infer/rules/common.rs
@@ -1,0 +1,137 @@
+use rustc_hash::{FxHashMap, FxHashSet};
+
+use crate::{
+    id::InFile, infer::pretty_print, scope::TypeConstructorKind, surface::tree::*,
+    InferenceDatabase,
+};
+
+use super::{CoreType, CoreTypeId, InferContext};
+
+impl InferContext<'_> {
+    pub(super) fn lower_type(&mut self, db: &dyn InferenceDatabase, type_id: TypeId) -> CoreTypeId {
+        match &self.arena[type_id] {
+            Type::Arrow(arguments, result) => {
+                let result = self.lower_type(db, *result);
+                let arguments = arguments.iter().map(|argument| self.lower_type(db, *argument));
+                arguments.rev().fold(result, |result, argument| {
+                    db.intern_type(CoreType::Function(argument, result))
+                })
+            }
+            Type::Application(function, arguments) => {
+                let function = self.lower_type(db, *function);
+                let arguments = arguments.iter().map(|argument| self.lower_type(db, *argument));
+                arguments.fold(function, |function, argument| {
+                    db.intern_type(CoreType::Application(function, argument))
+                })
+            }
+            Type::Constructor(_) => {
+                let resolution = self.resolve.per_type_type.get(&type_id);
+                db.intern_type(resolution.map_or(CoreType::NotImplemented, |resolution| {
+                    let file_id = resolution.file_id;
+                    match resolution.kind {
+                        TypeConstructorKind::Data(data_id) => {
+                            CoreType::Constructor(InFile { file_id, value: data_id })
+                        }
+                    }
+                }))
+            }
+            Type::Parenthesized(parenthesized) => self.lower_type(db, *parenthesized),
+            Type::Variable(name) => {
+                let name = Name::clone(name);
+                db.intern_type(CoreType::Variable(name))
+            }
+            Type::NotImplemented => db.intern_type(CoreType::NotImplemented),
+        }
+    }
+
+    pub(super) fn peel_arguments(
+        &mut self,
+        db: &dyn InferenceDatabase,
+        type_id: CoreTypeId,
+    ) -> Vec<CoreTypeId> {
+        let mut arguments = vec![];
+        let mut current = type_id;
+        while let CoreType::Function(argument, result) = db.lookup_intern_type(current) {
+            arguments.push(argument);
+            current = result;
+        }
+        arguments
+    }
+
+    pub(super) fn unify_types(&mut self, db: &dyn InferenceDatabase, x: CoreTypeId, y: CoreTypeId) {
+        println!("{} ~ {}", pretty_print(db, x), pretty_print(db, y));
+    }
+
+    pub(super) fn instantiate_type(
+        &mut self,
+        db: &dyn InferenceDatabase,
+        type_id: CoreTypeId,
+    ) -> CoreTypeId {
+        if let CoreType::Forall(initial_variable, initial_body) = db.lookup_intern_type(type_id) {
+            let mut replacements = FxHashMap::default();
+
+            let initial_unification = self.fresh_unification(db);
+            replacements.insert(initial_variable, initial_unification);
+            let mut current_body = initial_body;
+
+            while let CoreType::Forall(variable, body) = db.lookup_intern_type(current_body) {
+                let unification = self.fresh_unification(db);
+                replacements.insert(variable, unification);
+                current_body = body;
+            }
+
+            replace_type(db, &replacements, current_body)
+        } else {
+            type_id
+        }
+    }
+}
+
+fn replace_type(
+    db: &dyn InferenceDatabase,
+    replacements: &FxHashMap<Name, CoreTypeId>,
+    type_id: CoreTypeId,
+) -> CoreTypeId {
+    fn aux(
+        db: &dyn InferenceDatabase,
+        in_scope: &mut FxHashSet<Name>,
+        replacements: &FxHashMap<Name, CoreTypeId>,
+        type_id: CoreTypeId,
+    ) -> CoreTypeId {
+        match db.lookup_intern_type(type_id) {
+            CoreType::Application(function, argument) => {
+                let function = aux(db, in_scope, replacements, function);
+                let argument = aux(db, in_scope, replacements, argument);
+                db.intern_type(CoreType::Application(function, argument))
+            }
+            CoreType::Constructor(_) => type_id,
+            CoreType::Forall(variable, body) => {
+                in_scope.insert(Name::clone(&variable));
+                let body = aux(db, in_scope, replacements, body);
+                in_scope.remove(&variable);
+                db.intern_type(CoreType::Forall(variable, body))
+            }
+            CoreType::Function(argument, result) => {
+                let argument = aux(db, in_scope, replacements, argument);
+                let result = aux(db, in_scope, replacements, result);
+                db.intern_type(CoreType::Function(argument, result))
+            }
+            CoreType::Primitive(_) => type_id,
+            CoreType::Unification(_) => type_id,
+            CoreType::Variable(name) => {
+                if in_scope.contains(&name) {
+                    type_id
+                } else {
+                    replacements
+                        .get(&name)
+                        .copied()
+                        .unwrap_or_else(|| db.intern_type(CoreType::NotImplemented))
+                }
+            }
+            CoreType::NotImplemented => type_id,
+        }
+    }
+
+    let mut in_scope = FxHashSet::default();
+    aux(db, &mut in_scope, replacements, type_id)
+}
